requirements.txt
tensorflow
numpy
pandas
pillow
tqdm
matplotlib
